api/requirements.txt
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
gunicorn
whitenoise
dj-database-url
psycopg2-binary
python-dotenv
